+++ jsrelay.php
@@ -5,6 +5,7 @@ include "./config.php";
 include "./utils.php";
 
 $info["is_alive"] = is_alive($config);
+$info["last_heartbeat"] = last_heartbeat_age($config);
 
 echo json_encode($info);